@@ -4,7 +4,7 @@ extension PostgresMessage {
     /// Identifies an incoming or outgoing postgres message. Sent as the first byte, before the message size.
     /// Values are not unique across all identifiers, meaning some messages will require keeping state to identify.
     @available(*, deprecated, message: "Will be removed from public API.")
-    public struct Identifier: ExpressibleByIntegerLiteral, Equatable, CustomStringConvertible {
+    public struct Identifier: Sendable, ExpressibleByIntegerLiteral, Equatable, CustomStringConvertible {
         // special
         public static let none: Identifier = 0x00
         // special
