@@ -1,7 +1,7 @@
 import Logging
 import NIOCore
 
-enum HandlerTask {
+enum HandlerTask: Sendable {
     case extendedQuery(ExtendedQueryContext)
     case closeCommand(CloseCommandContext)
     case startListening(NotificationListener)
@@ -31,7 +31,7 @@ enum PSQLTask {
     }
 }
 
-final class ExtendedQueryContext {
+final class ExtendedQueryContext: Sendable {
     enum Query {
         case unnamed(PostgresQuery, EventLoopPromise<PSQLRowStream>)
         case executeStatement(PSQLExecuteStatement, EventLoopPromise<PSQLRowStream>)
@@ -100,14 +100,15 @@ final class PreparedStatementContext: Sendable {
     }
 }
 
-final class CloseCommandContext {
+final class CloseCommandContext: Sendable {
     let target: CloseTarget
     let logger: Logger
     let promise: EventLoopPromise<Void>
     
-    init(target: CloseTarget,
-         logger: Logger,
-         promise: EventLoopPromise<Void>
+    init(
+        target: CloseTarget,
+        logger: Logger,
+        promise: EventLoopPromise<Void>
     ) {
         self.target = target
         self.logger = logger
