@@ -1,4 +1,4 @@
-#if compiler(>=5.9)
+
 extension PostgresRow {
     // --- snip TODO: Remove once bug is fixed, that disallows tuples of one
     @inlinable
@@ -170,5 +170,3 @@ enum ComputeParameterPackLength {
         MemoryLayout<(repeat BoolConverter<each T>.Bool)>.size / MemoryLayout<Bool>.stride
     }
 }
-#endif // compiler(>=5.9)
-
